# backend/openapi.yaml
openapi: 3.0.3
info:
  title: AutoFlow Runtime API
  description: |
    REST API for the AutoFlow workflow automation platform.

    ## Overview
    AutoFlow exposes two primary resource families:
    - **Workflows** — named, versioned workflow definitions (analogous to templates). Supports full CRUD.
    - **Runs** — individual executions of a workflow. Runs are created by triggering a workflow and can be monitored for status, logs, and step-level traces.

    ## WebSocket Events
    Live run log streaming is available via WebSocket at `ws://<host>/api/runs/{runId}/events`.
    The server emits newline-delimited JSON event objects (see the `RunEvent` schema).
    Connect after creating a run; the stream closes automatically when the run reaches a terminal state (`completed`, `failed`).

    ## Current Implementation Notes
    - Run storage is currently in-memory. All runs are lost on server restart (see planned PostgreSQL migration).
    - Workflow CRUD endpoints (`POST/PUT/DELETE /workflows`) persist in-memory only in v0.
    - The `/api/templates/*` routes are legacy aliases maintained for backward compatibility.
  version: "1.0.0"
  contact:
    name: AutoFlow Backend Team
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.autoflow.example.com
    description: Production
...
paths:
...
  /api/llm-configs:
    post:
      operationId: createLlmConfig
      summary: Create an LLM provider config (store API key encrypted)
      tags: [LLMConfigs]
      parameters:
        - name: X-User-Id
          in: header
          required: true
          schema:
            type: string
          description: Authenticated user ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LLMConfigInput"
            example:
              provider: openai
              label: "My GPT-4o key"
              model: "gpt-4o"
              apiKey: "sk-..."
      responses:
        "201":
          description: Config created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LLMConfig"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

    get:
      operationId: listLlmConfigs
      summary: List user's LLM configs (keys masked)
      tags: [LLMConfigs]
      parameters:
        - name: X-User-Id
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of configs
          content:
            application/json:
              schema:
                type: object
                properties:
                  configs:
                    type: array
                    items:
                      $ref: "#/components/schemas/LLMConfig"
                  total:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/llm-configs/{id}:
    patch:
      operationId: updateLlmConfig
      summary: Update label or model
      tags: [LLMConfigs]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-User-Id
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                label:
                  type: string
                model:
                  type: string
      responses:
        "200":
          description: Updated config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LLMConfig"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      operationId: deleteLlmConfig
      summary: Remove a config
      tags: [LLMConfigs]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-User-Id
          in: header
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Config deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/llm-configs/{id}/default:
    patch:
      operationId: setLlmConfigDefault
      summary: Set as default provider for user's LLM steps
      tags: [LLMConfigs]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-User-Id
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated config (isDefault set to true)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LLMConfig"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------------------------------------------------------------------------
  # Legacy template aliases (backward-compatible)
  # ---------------------------------------------------------------------------
...
components:
...
  responses:
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "Workflow not found: lead-enrichment"
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "X-User-Id header is required"
  schemas:
...
    LLMProvider:
      type: string
      enum:
        - openai
        - anthropic
        - gemini
        - mistral
        - azure-openai
        - bedrock
        - vertex-ai
        - groq
        - fireworks
        - together
        - ollama
        - localai
        - cohere
        - perplexity
        - xai
        - deepseek
      description: Supported LLM provider
    LLMProviderCredentialsInput:
      type: object
      description: |
        Provider secret payload. Values are stored encrypted and never returned in plaintext.
        Use `apiKey` for API-key providers, AWS credentials for Bedrock, and either
        `serviceAccountJson` or `oauthAccessToken` for Vertex AI.
      properties:
        apiKey:
          type: string
          minLength: 4
        accessKeyId:
          type: string
        secretAccessKey:
          type: string
        sessionToken:
          type: string
        serviceAccountJson:
          type: string
        oauthAccessToken:
          type: string
    LLMProviderCredentialSummary:
      type: object
      description: Masked summary of stored secrets
      properties:
        apiKeyMasked:
          type: string
          example: "****abcd"
        accessKeyIdMasked:
          type: string
          example: "****MPLE"
        secretAccessKeyMasked:
          type: string
          example: "****KEY1"
        sessionTokenMasked:
          type: string
          example: "****9999"
        serviceAccountJsonMasked:
          type: string
          example: "****json"
        oauthAccessTokenMasked:
          type: string
          example: "****token"
    LLMProviderOptions:
      type: object
      description: Non-secret provider metadata
      properties:
        endpoint:
          type: string
          format: uri
          description: Custom provider endpoint or Azure OpenAI resource URL
        deployment:
          type: string
          description: Azure OpenAI deployment name
        apiVersion:
          type: string
          description: Provider API version metadata, such as Azure OpenAI api-version
        region:
          type: string
          description: AWS region for Bedrock credentials
        projectId:
          type: string
          description: GCP project ID for Vertex AI
        location:
          type: string
          description: GCP location/region for Vertex AI
        authType:
          type: string
          enum: [api_key, aws, service_account, oauth]
    LLMConfigInput:
      type: object
      required: [provider, label, model]
      properties:
        provider:
          $ref: "#/components/schemas/LLMProvider"
        label:
          type: string
          description: User-friendly name for this config
          example: "My GPT-4o key"
        model:
          type: string
          description: Model identifier for the provider
          example: "gpt-4o"
        apiKey:
          type: string
          description: |
            Backward-compatible shorthand for API-key providers. Stored encrypted and
            never returned in plaintext.
          example: "sk-..."
          minLength: 4
          deprecated: true
        credentials:
          $ref: "#/components/schemas/LLMProviderCredentialsInput"
        providerOptions:
          $ref: "#/components/schemas/LLMProviderOptions"
    LLMConfig:
      type: object
      required: [id, userId, provider, label, model, credentialSummary, isDefault, createdAt]
      description: LLM provider config with encrypted secret storage and masked secret summaries
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          description: Owner user ID
        provider:
          $ref: "#/components/schemas/LLMProvider"
        label:
          type: string
          example: "My GPT-4o key"
        model:
          type: string
          example: "gpt-4o"
        apiKeyMasked:
          type: string
          description: Backward-compatible masked API key field for legacy clients
          example: "****abcd"
        credentialSummary:
          $ref: "#/components/schemas/LLMProviderCredentialSummary"
        providerOptions:
          $ref: "#/components/schemas/LLMProviderOptions"
        isDefault:
          type: boolean
          description: Whether this is the user's default LLM config
        createdAt:
          type: string
          format: date-time

    # -------------------------------------------------------------------------
    # Error
    # -------------------------------------------------------------------------
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
          example: "Workflow not found: lead-enrichment"